+++ src/osd/scheduler/mClockScheduler.h
@@ -26,73 +26,12 @@
 
 #include "osd/scheduler/OpScheduler.h"
 #include "common/config.h"
+#include "common/mclock_common.h"
 #include "common/ceph_context.h"
 #include "osd/scheduler/OpSchedulerItem.h"
 
 
-enum {
-  l_mclock_first = 15000,
-  l_mclock_immediate_queue_len,
-  l_mclock_client_queue_len,
-  l_mclock_recovery_queue_len,
-  l_mclock_best_effort_queue_len,
-  l_mclock_all_type_queue_len,
-  l_mclock_last,
-};
-
 namespace ceph::osd::scheduler {
-
-constexpr double default_min = 0.0;
-constexpr double default_max = std::numeric_limits<double>::is_iec559 ?
-  std::numeric_limits<double>::infinity() :
-  std::numeric_limits<double>::max();
-
-/**
- * client_profile_id_t
- *
- * client_id - global id (client.####) for client QoS
- * profile_id - id generated by client's QoS profile
- *
- * Currently (Reef and below), both members are set to
- * 0 which ensures that all external clients share the
- * mClock profile allocated reservation and limit
- * bandwidth.
- *
- * Note: Post Reef, both members will be set to non-zero
- * values when the distributed feature of the mClock
- * algorithm is utilized.
- */
-struct client_profile_id_t {
-  uint64_t client_id = 0;
-  uint64_t profile_id = 0;
-
-  client_profile_id_t(uint64_t _client_id, uint64_t _profile_id) :
-    client_id(_client_id),
-    profile_id(_profile_id) {}
-
-  client_profile_id_t() = default;
-
-  auto operator<=>(const client_profile_id_t&) const = default;
-  friend std::ostream& operator<<(std::ostream& out,
-                                  const client_profile_id_t& client_profile) {
-    out << " client_id: " << client_profile.client_id
-        << " profile_id: " << client_profile.profile_id;
-    return out;
-  }
-};
-
-struct scheduler_id_t {
-  op_scheduler_class class_id;
-  client_profile_id_t client_profile_id;
-
-  auto operator<=>(const scheduler_id_t&) const = default;
-  friend std::ostream& operator<<(std::ostream& out,
-                                  const scheduler_id_t& sched_id) {
-    out << "{ class_id: " << sched_id.class_id
-        << sched_id.client_profile_id;
-    return out << " }";
-  }
-};
 
 /**
  * Scheduler implementation based on mclock.
@@ -102,87 +41,11 @@ struct scheduler_id_t {
 class mClockScheduler : public OpScheduler, md_config_obs_t {
 
   CephContext *cct;
-  const int whoami;
-  const uint32_t num_shards;
-  const int shard_id;
-  const bool is_rotational;
   const unsigned cutoff_priority;
   MonClient *monc;
-  PerfCounters *logger;
 
-  /**
-   * osd_bandwidth_cost_per_io
-   *
-   * mClock expects all queued items to have a uniform expression of
-   * "cost".  However, IO devices generally have quite different capacity
-   * for sequential IO vs small random IO.  This implementation handles this
-   * by expressing all costs as a number of sequential bytes written adding
-   * additional cost for each random IO equal to osd_bandwidth_cost_per_io.
-   *
-   * Thus, an IO operation requiring a total of <size> bytes to be written
-   * accross <iops> different locations will have a cost of
-   * <size> + (osd_bandwidth_cost_per_io * <iops>) bytes.
-   *
-   * Set in set_osd_capacity_params_from_config in the constructor and upon
-   * config change.
-   *
-   * Has units bytes/io.
-   */
-  double osd_bandwidth_cost_per_io;
-
-  /**
-   * osd_bandwidth_capacity_per_shard
-   *
-   * mClock expects reservation and limit paramters to be expressed in units
-   * of cost/second -- which means bytes/second for this implementation.
-   *
-   * Rather than expecting users to compute appropriate limit and reservation
-   * values for each class of OSDs in their cluster, we instead express
-   * reservation and limit paramaters as ratios of the OSD's maxmimum capacity.
-   * osd_bandwidth_capacity_per_shard is that capacity divided by the number
-   * of shards.
-   *
-   * Set in set_osd_capacity_params_from_config in the constructor and upon
-   * config change.
-   *
-   * This value gets passed to ClientRegistry::update_from_config in order
-   * to resolve the full reservaiton and limit parameters for mclock from
-   * the configured ratios.
-   *
-   * Has units bytes/second.
-   */
-  double osd_bandwidth_capacity_per_shard;
-
-  class ClientRegistry {
-    std::array<
-      crimson::dmclock::ClientInfo,
-      static_cast<size_t>(op_scheduler_class::immediate)
-    > internal_client_infos = {
-      // Placeholder, gets replaced with configured values
-      crimson::dmclock::ClientInfo(1, 1, 1),
-      crimson::dmclock::ClientInfo(1, 1, 1)
-    };
-
-    crimson::dmclock::ClientInfo default_external_client_info = {1, 1, 1};
-    std::map<client_profile_id_t,
-	     crimson::dmclock::ClientInfo> external_client_infos;
-    const crimson::dmclock::ClientInfo *get_external_client(
-      const client_profile_id_t &client) const;
-  public:
-    /**
-     * update_from_config
-     *
-     * Sets the mclock paramaters (reservation, weight, and limit)
-     * for each class of IO (background_recovery, background_best_effort,
-     * and client).
-     */
-    void update_from_config(
-      const ConfigProxy &conf,
-      double capacity_per_shard);
-    const crimson::dmclock::ClientInfo *get_info(
-      const scheduler_id_t &id) const;
-  } client_registry;
-
+  ClientRegistry client_registry;
+  MclockConfig mclock_conf;
   using mclock_queue_t = crimson::dmclock::PullPriorityQueue<
     scheduler_id_t,
     OpSchedulerItem,
@@ -210,23 +73,6 @@ class mClockScheduler : public OpScheduler, md_config_obs_t {
     };
   }
 
-  /**
-   * set_osd_capacity_params_from_config
-   *
-   * mClockScheduler uses two parameters, osd_bandwidth_cost_per_io
-   * and osd_bandwidth_capacity_per_shard, internally.  These two
-   * parameters are derived from config parameters
-   * osd_mclock_max_capacity_iops_(hdd|ssd) and
-   * osd_mclock_max_sequential_bandwidth_(hdd|ssd) as well as num_shards.
-   * Invoking set_osd_capacity_params_from_config() resets those derived
-   * params based on the current config and should be invoked any time they
-   * are modified as well as in the constructor.  See handle_conf_change().
-   */
-  void set_osd_capacity_params_from_config();
-
-  // Set the mclock related config params based on the profile
-  void set_config_defaults_from_profile();
-
 public: 
   template<typename Rep, typename Per>
   mClockScheduler(
@@ -238,15 +84,12 @@ public:
     MonClient *monc,
     bool init_perfcounter=true)
     : cct(cct),
-      whoami(whoami),
-      num_shards(num_shards),
-      shard_id(shard_id),
-      is_rotational(is_rotational),
       cutoff_priority(cutoff_priority),
       monc(monc),
-      logger(nullptr),
+      mclock_conf(cct, client_registry, monc, num_shards,
+	          is_rotational, shard_id, whoami),
       scheduler(
-	std::bind(&mClockScheduler::ClientRegistry::get_info,
+	std::bind(&ClientRegistry::get_info,
 		  &client_registry,
 		  std::placeholders::_1),
 	idle_age, erase_age, check_time,
@@ -255,13 +98,13 @@ public:
   {
     cct->_conf.add_observer(this);
     ceph_assert(num_shards > 0);
-    set_osd_capacity_params_from_config();
-    set_config_defaults_from_profile();
+    mclock_conf.set_osd_capacity_params_from_config();
+    mclock_conf.set_config_defaults_from_profile();
     client_registry.update_from_config(
-      cct->_conf, osd_bandwidth_capacity_per_shard);
+      cct->_conf, mclock_conf.get_capacity_per_shard());
 
     if (init_perfcounter) {
-      _init_logger();
+      mclock_conf.init_logger();
     }
   }
   mClockScheduler(
@@ -318,14 +161,11 @@ public:
 			  const std::set<std::string> &changed) final;
 
   double get_cost_per_io() const {
-    return osd_bandwidth_cost_per_io;
+    return mclock_conf.get_cost_per_io();
   }
 private:
   // Enqueue the op to the high priority queue
   void enqueue_high(unsigned prio, OpSchedulerItem &&item, bool front = false);
-  void _init_logger();
-  void _get_mclock_counter(scheduler_id_t id);
-  void _put_mclock_counter(scheduler_id_t id);
 };
 
 }
